Move ssh tunnel to a new file
